+++ src/main/java/io/github/a5b84/helditeminfo/mixin/InGameHudMixin.java
@@ -88,7 +88,7 @@ public abstract class InGameHudMixin {
             context.fill(
                     (scaledWidth - maxWidth) / 2 - 2, y - 2,
                     (scaledWidth + maxWidth) / 2 + 2, y + height + 2,
-                    color);
+                    backgroundColor);
         }
 
         int lineHeight = config.lineHeight();