+++ src/main/java/io/github/a5b84/helditeminfo/mixin/item/BannerPatternItemMixin.java
@@ -13,5 +13,4 @@ public abstract class BannerPatternItemMixin implements GenericTooltipAppender {
     public boolean heldItemInfo_shouldAppendTooltip() {
         return config.showBannerPatternName();
     }
-
 }