+++ src/main/java/io/github/a5b84/helditeminfo/mixin/item/CrossbowItemMixin.java
@@ -24,5 +24,4 @@ public abstract class CrossbowItemMixin implements GenericTooltipAppender {
         Util.setAllToDefaultColor(tooltip);
         return tooltip;
     }
-
 }