+++ src/main/java/io/github/a5b84/helditeminfo/mixin/item/DiscFragmentItemMixin.java
@@ -13,5 +13,4 @@ public abstract class DiscFragmentItemMixin implements GenericTooltipAppender {
     public boolean heldItemInfo_shouldAppendTooltip() {
         return config.showMusicDiscDescription();
     }
-
 }