+++ src/main/java/io/github/a5b84/helditeminfo/mixin/item/EntityBucketItemMixin.java
@@ -13,5 +13,4 @@ public abstract class EntityBucketItemMixin implements GenericTooltipAppender {
     public boolean heldItemInfo_shouldAppendTooltip() {
         return config.showEntityBucketContent();
     }
-
 }