+++ src/main/java/io/github/a5b84/helditeminfo/mixin/item/FireworkItemsMixin.java
@@ -14,5 +14,4 @@ public abstract class FireworkItemsMixin implements GenericTooltipAppender {
     public boolean heldItemInfo_shouldAppendTooltip() {
         return config.showFireworkAttributes();
     }
-
 }