+++ src/main/java/io/github/a5b84/helditeminfo/mixin/item/ItemsWithPotionEffectsMixin.java
@@ -34,5 +34,4 @@ public abstract class ItemsWithPotionEffectsMixin implements GenericTooltipAppen
 
         return tooltip;
     }
-
 }